Disposable

+++ StartApp/Program.cs
@@ -9,13 +9,14 @@ namespace StartApp
         static void Main(string[] args)
         {
             var workingDirectory = @"D:\Projects\SimpleDB\StartApp\bin\Debug\netcoreapp3.1\Database";
-            foreach (var file in Directory.GetFiles(workingDirectory))
+            var modeCreate = false;
+
+            if (modeCreate)
             {
-                File.Delete(file);
+                foreach (var file in Directory.GetFiles(workingDirectory)) File.Delete(file);
             }
 
             // build engine
-
             var builder = DBEngineBuilder.Make();
             builder.WorkingDirectory(workingDirectory);
             builder.Map<Person>()
@@ -27,22 +28,25 @@ namespace StartApp
                 .Field(3, x => x.AdditionalInfo);
 
             var engine = builder.BuildEngine();
-
-            // insert
-
-            Console.WriteLine("========== Insert ==========");
             var collection = engine.GetCollection<Person>();
-            var sw = System.Diagnostics.Stopwatch.StartNew();
+
             var count = 10000;
-            for (int i = 0; i < count; i++)
+            System.Diagnostics.Stopwatch sw = null;
+
+            if (modeCreate)
             {
-                collection.Insert(new Person { Id = i, Name = "Name " + i, Surname = "Surname " + i, Middlename = "Middlename " + i, AdditionalInfo = new PersonAdditionalInfo { Value = i } });
+                // insert
+                Console.WriteLine("========== Insert ==========");
+                sw = System.Diagnostics.Stopwatch.StartNew();
+                for (int i = 0; i < count; i++)
+                {
+                    collection.Insert(new Person { Id = i, Name = "Name " + i, Surname = "Surname " + i, Middlename = "Middlename " + i, AdditionalInfo = new PersonAdditionalInfo { Value = i } });
+                }
+                sw.Stop();
+                Console.WriteLine(sw.Elapsed);
             }
-            sw.Stop();
-            Console.WriteLine(sw.Elapsed);
 
             // get
-
             Console.WriteLine("========== Get ==========");
             sw = System.Diagnostics.Stopwatch.StartNew();
             for (int i = 0; i < count; i++)
@@ -58,12 +62,18 @@ namespace StartApp
             Console.WriteLine(collection.Get(count - 1));
 
             // update
-
             Console.WriteLine("========== Update ==========");
             sw = System.Diagnostics.Stopwatch.StartNew();
             for (int i = 0; i < count; i++)
             {
-                collection.Update(new Person { Id = i, Name = "New Name " + i, Surname = "New Surname " + i, Middlename = "New Middlename " + i, AdditionalInfo = new PersonAdditionalInfo { Value = -i } });
+                if (collection.Exist(i))
+                {
+                    collection.Update(new Person { Id = i, Name = "New Name " + i, Surname = "New Surname " + i, Middlename = "New Middlename " + i, AdditionalInfo = new PersonAdditionalInfo { Value = -i } });
+                }
+                else
+                {
+                    Console.WriteLine(String.Format("Id {0} not exists", i));
+                }
             }
             sw.Stop();
             Console.WriteLine(sw.Elapsed);
@@ -74,29 +84,45 @@ namespace StartApp
             Console.WriteLine(collection.Get(count - 1));
 
             // delete
-
             Console.WriteLine("========== Delete ==========");
-            collection.Delete(0);
+            if (collection.Exist(0))
+            {
+                collection.Delete(0);
+            }
+            else
+            {
+                Console.WriteLine(String.Format("Id {0} not exists", 0));
+            }
             Console.WriteLine(collection.Get(0) == null ? "collection.Get(0): null" : "collection.Get(0): !!! not null !!!");
             Console.WriteLine("collection.Get(1): " + collection.Get(1));
 
             // linq query
-
             Console.WriteLine("========== Linq query ==========");
 
             var queryResult = collection.Query()
                 .Select(x => new { x.Id, x.Name, x.AdditionalInfo })
                 .Where(x => x.Name == "New Name 10" && x.Surname == "New Surname 10")
-                .OrderBy(x => x.Name, SortDirection.Asc)
-                .OrderBy(x => x.Surname, SortDirection.Desc)
-                .Skip(1)
+                .ToList();
+            foreach (var item in queryResult) Console.WriteLine(item);
+            Console.WriteLine("- - - - - - - - -");
+
+            queryResult = collection.Query()
+                .Select(x => new { x.Id, x.Name })
+                .Skip(10)
                 .Limit(10)
                 .ToList();
+            foreach (var item in queryResult) Console.WriteLine(item);
+            Console.WriteLine("- - - - - - - - -");
 
-            foreach (var item in queryResult)
-            {
-                Console.WriteLine(item);
-            }
+            queryResult = collection.Query()
+                .Select(x => new { x.Id, x.Name })
+                .Where(x => x.Name.Contains("111"))
+                .OrderBy(x => x.Id, SortDirection.Desc)
+                .ToList();
+            foreach (var item in queryResult) Console.WriteLine(item);
+            Console.WriteLine("- - - - - - - - -");
+
+            engine.Dispose();
 
             Console.ReadKey();
         }
@@ -116,7 +142,7 @@ namespace StartApp
 
         public override string ToString()
         {
-            return $"{Id}: {Name} {Surname} {Middlename} {AdditionalInfo.Value}";
+            return String.Format("{0}:\t{1}\t{2}\t{3}\t{4}", Id, Name, Surname, Middlename, AdditionalInfo != null ? AdditionalInfo.Value.ToString() : "null");
         }
     }
